Expand and blacklist a known failure in GregorianCalendarTest.
Expands the test to make the expectations clearer. The
issue here relates to how we deal with ambiguous wall clock
times, i.e times that occur during the time that's lost when
transitioning into daylight savings, or the wall clock
time that occurs twice when we transition out of daylight
savings.

NOTE: This isn't a regression, we've never handled this
correctly, despite a proud comment near the end of computeTime
that claims we've implemented a  "trick" to get this right.

bug: 12778197
Change-Id: Ida79583c02c6895096a8afdd8f59b1002b282d32