CTS libcore test:Fix test_keySet seldom fail
This test is passed with an assumption; GC can be completed within five loop counts.
But in fact, call System.gc() will not execute GC immediately. It depends on system loading and GC policy. We can reproduce this fail with a large number of repeat this test case.
So increasing the above loop counts.
Signed-off-by: Junjie.Hu <junjie.hu@mediatek.com>

Change-Id: I053513b007b00e627620f4bce3840493f8c3b279