CTS libcore test:Fix testServerShutdownOutput seldom fail
This test is passed with an assumption, which is server.getUrl("/a") is been
executed before server.getUrl("/b").But in real world, they are different
threads. If server.getUrl("/b") is been executed before server.getUrl("/a"),
then the response is not the same with expected.
So add a delay before run server.getUrl("/b).

Change-Id: I41134c9ef24d69df8358e5177df7170ab76f8fde
Signed-off-by: Junjie.Hu <junjie.hu@mediatek.com>