libcore: Clean up Enum.compareTo(E)
As E extends Enum<E>, it is safe to cast the comparable object to
Enum<?> as it should not throw any exceptions. When it is cast, the
ordinal value can be accessed directly. Doing so would be preferred as
either both the variables this.ordinal and o.ordinal should be used or
the methods this.ordinal() and o.ordinal() should be used without
mixing them together.

Change-Id: I7c4e93fdf35b0bf2949738a91c0855d9a52e5c13