Use the provided path when opening dex files.
Currently we pass down to the native opener the canonical path.
If the path contains symlinks it may generate the wrong file name
when searching for the containing oat file. This is because we
don't apply any canonicalization when dex files are generated so
we think we're looking for somthing else.

Bug: 15563230
Change-Id: I13c84eace2320d84d530195407bed91d20c0c342