Fix Harmony-707 test for CTS
Change the way Harmony-707 test works.
We cannot check timeouts equality, as timeout is approximated by the Linux Kernel (see set_sock_timeout() in net/core/sock.c).
Instead, set 2 different timeout and ensure they are not equals after calling accept() two times.

Testing only with one value is not the right way to do it:
- On a device running a kernel compiled with CONFIG_HZ=100 or 1000, it will work
- On a device running a kernel compiled with CONFIG_HZ=250, it wont:
  * Kernel will store a timeout of 3 (3.49975 cast to long)
   *timeo_p = tv.tv_sec*HZ + (tv.tv_usec+(1000000/HZ-1))/(1000000/HZ)
   with tv.tv_sec=0, tv.tv_usec=10000 (10ms), and HZ=250
  * Kernel will return a timeout of 12ms
   v.tm.tv_usec = ((sk->sk_rcvtimeo % HZ) * 1000000) / HZ
   with sk->sk_rcvtimeo=3 and HZ=250

This patch is a new way for testing Harmony-707, by setting 2 timeouts and ensure they are not both set to the default timeout value.

Change-Id: I7f05c6843d9e48e6d696dca17173fce028a45a68
Signed-off-by: Samuel Rats <samuel.rats@gmail.com>