Revert to old behaviour for callers LocaleData.get(null).
Some of our APIs were (incorrectly IMO) falling back to
the default locale when presented with a null locale because
LocaleData.get(null) did that behind the scenes.

We changed that behaviour, so we have to go change all
possible callers of this API.

Also, this change makes the NPE for null locales explicit
in places that would already throw NPE for other reasons.

bug: 16938922
Change-Id: I3e77435a6b414b002b96a4f8deb44b51591c5a51