Assert default SSLSocketFactory and HostnameVerifier actually used.
This adds tests to assert that the default SSLSocketFactory and
HostnameVerifier of HttpsURLConnection is actually used for new
instances of HttpsURLConnection.

The contract of HttpsURLConnection.get/setDefaultHostnameVerifier
and HttpsURLConnection.get/setDefaultSSLSocketFactory is that the
HostnameVerifier and SSLSocketFactory returned/set via these
methods are the ones used for new instances of HttpsURLConnection.

Bug: 18481199
Change-Id: Id976c2fc31f0e3efed9284860aa35144a729dabb