Move endif statement to enclose all related rules.
The rules to build a resource file with a "#" in its name on-the-fly
apply to the preceding BUILD_STATIC_JAVA_LIBRARY rules and don't do
anything useful if those rules are skipped because LIBCORE_SKIP_TESTS is
set. While it doesn't cause any harm to have these rules there anyway,
it'll become clearer to the reader if they're included in the if-endif
block.

Change-Id: Iddc76e994e00c38c5733a30e6d039480a5f557b2