CipherSpi.engineDoFinal does not permit null input buffer.
javax.crypto.Cipher.doFinal() and Cipher.doFinal(byte[] output,
int outputOffset) invokes the underlying CipherSpi.engineDoFinal with
null as the input buffer. This is against the contract of
CipherSpi.engineDoFinal which does not permit null for the input
buffer argument.

This CL fixes the implementation of Cipher to use an empty byte array
instead of null in these invocations of CipherSpi.engineDoFinal.

Bug: 19099415
Change-Id: I1854ffc5c073f84fbd070a32c9b68ea80892ce67