Make Cipher.update return null when no output produced.
The contract of Cipher.update for block ciphers is that it returns
null when the input was insufficient to produce a new block of
output. The current implementation does not enforce this and may
return an empty byte array instead, because it leaves it up to the
underlying CipherSpi to honor the contract.

This CL fixes the issue by making the Cipher implementation enforce
this contract.

Bug: 19100268
Change-Id: Ib7053a6ae06e1faef4cdde86343a7625a26381a4