Timezone data installer code
The code here is used by the system server to install timezone data updates.
It is separate so it can be tested.

Scripts are included that build an "update bundle" (a zip file with a
well-defined contents).

ConfigBundle contains logic to extract the bundle safely.
TzDataBundleBuilder is used in the test and tools to construct a bundle.

The script in tools is a placeholder and will evolve.
bionic/libc/tools/zoneinfo tools will likely move there so they can be
integrated more closely.

An app is included for testing updates.

Bug: 10637303
Change-Id: Id0985f8c5be2f12858ee8bf52acf52bdb2df8741