Fix the year display when formating absolute date string
In getRelativeDateTimeString(), when formatting the absolute date string
with formatDateRange(), we need to explicitly specify the option to
show / hide the year display based on the given arguments. Otherwise it
would determine based on the current time and may gave wrong results.

In getRelativeTimeSpanString(), it follows the same logic when user
doesn't supply any option.

Bug: 19822016
Change-Id: I0d4f01b7d15e0b8ea44f21b77309b1fc370b365b