Don't use zip64 by default.
This change was breaking tools like dexmaker that would write
zipfiles to disk so that the runtime could read them.

We don't want to go back and amend the local file header, so
we just throw an IOException if somebody writes more than 4GB
of data to an entry that doesn't have a declared size. If the
entry has a declared size > 4GB, we can write the zip64 header
in advance.

bug: 19574093
Change-Id: I771e821996efd3ef4c3de5e02f5a0062e62f9244