Add icu4j to core-libart.jar.
The first approach of building icu4j against the SDK and adding that
to the boot classpath won't work out too well. Besides being a bit
of a hack, there's no way to express that relationship for hostdex
builds.

We now compile icu4j sources along with core-libart. This is the most
direct way of expressing this relationship and allows us to include
exactly what we need.

Change-Id: Idd305e77480bc681ed5b47e740dfec20d3bc7b26