Small refactoring changes to System before making ICU4J changes
Updates to documentation around lineSeparator.
Motivation: To make PATH_SEPARATOR a constant so it can be used
when creating a compound ICU4J path. lineSeparator looks odd by
comparison and we can save a small amount of static
initialization by making it a constant now that line.separator
is immutable.

Change-Id: I20d17b6337a6cc420a057e5ec36b39aae8c3cb8a