Avoid using dex cache array pointers in libart.
In preparation for making dex cache arrays native, avoid
using them in Java code.

This causes a performance regression for our reflection
benchmarks. To reduce that regression, we add an intrinsic
for Class.findOverriddenMethodIfProxy(). Without that, the
Class_getDeclaredMethod and Class_getMethod benchmarks take
an up to 30% hit, with the intrinsic it's about 15%.
(Measured using the Quick compiler, the intrinsic isn't
implemented for Optimizing.)

Companion art/ change:
    https://android-review.googlesource.com/146068

Bug: 20134538
Change-Id: Ibbef3b50043a1311cd40723ed42e1f1c609b8fc1