Avoid using dex cache array pointers in libart.
In preparation for making dex cache arrays native, avoid
using them in Java code.

This causes a performance regression for our reflection
benchmarks. Class_getDeclaredMethod and Class_getMethod
take an up to 30% hit, measured using the Quick compiler.
We accept this hit at this stage and will tune the
performance after the we're done with the larger effort.

Companion art/ change:
    https://android-review.googlesource.com/146068

Bug: 20134538
Change-Id: Ibbef3b50043a1311cd40723ed42e1f1c609b8fc1