Migrate TimeZoneNames to ICU4J
TimeZoneNames was identified as one of the classes that
takes a long time to initialize during zygote init.

This change switches us to an equivalent ICU4J implementation.
Only one intended logic change: ICU's getDisplayName() was
being called in all cases. The isUtc() check was moved to
bypass these. The rest should be a straight port.

A benchmark was written but requires public access to
TimeZoneNames.createZoneStrings() so will not be checked in.

Results on AOSP/master with a mako, results in millis:

Before:
Fr_native 20.4
Uk_native 20.6

After:
Fr 10.6
Uk 10.6

Change-Id: Ibaa969cb477bd9b8cd52c6cbe629f1b33dca5e40