Migrate TimeZoneNames to ICU4J
This change switches us from an ICU4C to an equivalent
ICU4J implementation.

A benchmark was written but requires public access to
TimeZoneNames.createZoneStrings() so will not be checked in.

Results on AOSP/master with a hammerhead, results in millis:

Before:
Fr 12.4
Uk 12.7

After:
Fr 6.16
Uk 6.03

The static initializer now does some work up-front to obtain
and stash the canonical ID for each timezone. This was to avoid
the potential expense of getCanonicalID() each time
createZoneStrings() is called.  Although ICU4J caches the lookup
it uses a soft-reference cache and so it is not reliable or
deterministic. Adding the getCanonicalID() lookup shouldn't cost
any additional time but will take up some more permanent memory.

Change-Id: Ibaa969cb477bd9b8cd52c6cbe629f1b33dca5e40