KeyPairGeneratorTest: recognise that the OpenSSL provider won't parse every ECC curve.
Since the switch to BoringSSL, the OpenSSL provider won't parse every
possible ECC curve (for example, prime239v1, which has, most likely,
never been used for anything!)

However, KeyPairGeneratorTest will generate everything that BouncyCastle
knows about and try to get the OpenSSL provider to parse it. This change
alters the test to detect known OIDs in the encoded keys generated by
BouncyCastle and assert that only these OIDs are parsed by the OpenSSL
provider.

In order to avoid the complexity of an ASN.1 parser, the OID matching is
done via a substring match. That's a little ugly but the structure of
the PKCS#8 data should allow that to work without being flakey.

Bug: 21211314
Change-Id: I88063ca586a07b570acb5ffddb0ce87b4a6672fd