Libcore: workaround for Math.sqrt() when run x86 host interpreter
This is a workaround for inaccurate calculations produced by
Math.sqrt() in x86 host mode. Math.sqrt() calls sqrt() from
glibc libm.so. In some cases x86 sqrt() produces results
different from x86_64 sqrt() when run in host mode. It also
produces results different from x86 target mode. Such a behavior
is unallowable because results must be the same regardless of
mode. We may use StrictMath.sqrt() which is supposed to be more
accurate. It is slower than Math.sqrt() but we may afford it
because this code will be executed only in the interpreter mode.
In the compiler mode Math.sqrt() is always inlined.

Change-Id: Ic760a70b50e1fab3e1f99f41b8217d49a8ca1109
Reviewed-by: Serdjuk, Nikolay Y <nikolay.y.serdjuk@intel.com>