Optimizing String.Equals as a native method
This is the first of several potential solutions to optimizing
String.Equals. I added a JNI transition and called the native
memcmp function, very similar to how compareTo is currently
implemented. The goal is to benchmark this as well as several
other solutions to find the fastest implementation (or
combination of implementations) for this function.

Interesting Benchmarking Values:

Quick Compiler on Nexus 7
	JNI 15-30 Character Strings = 864.67 ns
	Original 15-30 Character Strings = 217.67 ns
	JNI 1000+ Character Strings = 17855.33 ns
	Original 1000+ Character Strings = 29328.67 ns

Quick Compiler on 64 bit Emulator
        JNI 15-30 Character Strings = 901 ns
        Original 15-30 Character Strings = 963.67 ns
        JNI 1000+ Character Strings = 1303 ns
        Original 1000+ Character Strings = 5766.67 ns

Bug: 21481923
Change-Id: I200305db2075d103b905083533de16df84c0a2aa