Optimizing String.Equals as a native method
Work In Progress. This is the first of several potential solutions to optimizing
String.Equals. I added a JNI transition and called the native memcmp function,
very similar to how compareTo is currently implemented. The goal is to benchmark
this as well as several other solutions to find the fastest implementation (or
combination of implementations) for this function.

Bug: 21481923
Change-Id: I200305db2075d103b905083533de16df84c0a2aa