Migrate DecimalFormat to use ICU4J.
This change migrates the underlying implementation of
DecimalFormat to use the Java version of the ICU library.

All all issues related to Harmony test failures have been
tracked down and everything not yet fixed in ICU has a
draft fix in the following CLs:

https://android-review.googlesource.com/#/c/173661/
https://android-review.googlesource.com/#/c/173662/
https://android-review.googlesource.com/#/c/173663/

To verify that all tests can pass with a fix in ICU code,
please patch in the above CL, and then run the tests.

Behaviour change: Now a new instance of DecimalFormat()
has 309 as the maximumIntegerDigits instead of 20000000
which it was before. This is because it doesn't make
sense to have this value when ICU will only use up to 309
to format numbers.

Furthermore, ICU4J now returns patterns in the form of
<positive-part>;<negative-part> for displaying the
difference between positive and negative number formats.
Tests have been adjusted accordingly.

Benchmarks:
                            benchmark    JAVA  NATIVE
              _formatAsExponent10e100   79.40    39.3
             _formatAsExponent10e1000   10.78    33.4
                _formatAsExponent10e3   33.10    32.2
                _formatAsExponent10e9   34.65    32.8
              _formatBigDecimal10e100  109.00    71.4
             _formatBigDecimal10e1000  125.23   308.1
                _formatBigDecimal10e3   34.58    28.9
                _formatBigDecimal10e9   34.72    33.7
                             _formatE   12.35    38.9
                           _formatEur    6.50    22.0
                  _formatEurWithCents   10.43    37.0
     _formatGrouping_BigDecimal10e100  108.01    71.5
    _formatGrouping_BigDecimal10e1000  126.39   313.2
       _formatGrouping_BigDecimal10e3   33.92    29.2
       _formatGrouping_BigDecimal10e9   34.91    33.6
                            _formatPi   12.46    38.7
     _formatToCharacterIterator10e100  629.03   153.3
    _formatToCharacterIterator10e1000 3200.23  3906.3
       _formatToCharacterIterator10e3   66.99    34.0
       _formatToCharacterIterator10e9   87.77    44.2
                           _formatUSD    7.20    30.1
                  _formatUsdWithCents   10.55    37.9

Tested on Nexus 6.

Change-Id: I642fd02acd3bcd198960cdf2a8d973db7baec42b