Migrate the Transliterator to ICU4J.
This change removes the native bindings to the ICU4C transliterator,
and calls the ICU4J transliterator instead.

A benchmark was performed which shows the performance impact of
this change.

                benchmark    us linear runtime
       javaTransliterator 206.4 ==============================
     nativeTransliterator  46.9 ======
    vm: app_process

Additionally this change requires the addition of the ICU4J
transliterator code into the packaged jar, which increases in size
from 2.289MB before the addition, to 2.332 afterwards.

Change-Id: I45a031c98ee9fbc027cc19dcdc70682b9d7dd6e9