(WIP) Hybrid Implementation of String.Equals
The final implementation of String.Equals, splitting the method into a
java method which includes check like instanceof and reference equality
which could potentially be optimized out by the compiler, and an
intrisified helper function to do the character comparison several
characters at a time.

I'm still working on this- for example, I would like to add count as an
argument to the loop helper function so I can avoid a load instruction in
my intrinsic. I'm open to any suggestions about how to improve this more.

Bug: 21481923
Change-Id: I94a30a39da307cd7b699eccf5e2110ebb83a2edf