Add ability to prevent applications from overriding TrustManager
Add allowTLSTrustManagerOverride variable and methods to
NetworkSecurityPolicy. The variable defaults to true. When the
variable is false, then the platform's TLS library will always
use the platform default TrustManager even when the application
specifies a different TrustManager. This is a best effort mechanism
that is only effective for applications that use the Android built-in
TLS library rather than bringing along their own, although
third-party implementations are encouraged to also honor this flag.

As described by numerous sources including [1] [2] [3] and [4], custom
TrustManagers are a common source of security vulnerabilities.

This change goes along with corresponding changes to external/conscrypt
and frameworks/base with the same Change-Id.

[1] Tendulkar and Enck. An Application Package Configuration Approach
to Mitigating Android SSL Vulnerabilities. MoST 2014.
[2] Fahl, et al. Why Eve and Mallory Love Android: An Analysis of
Android SSL (In)Security. CCS '12.
[3] Sounthiraraj, et al. Large Scale, Automated Detection of SSL/TLS
Man-in-the-Middle Vulnerabilities in Android Apps. NDSS '14.
[4] FireEye. SSL Vulnerabilities: Who listens when Android
applications talk?
https://www.fireeye.com/blog/threat-research/2014/08/ssl-vulnerabilities-who-listens-when-android-applications-talk.html

Change-Id: If6ab0ab309488926cdde7fd55539bd059eea964d