Add ability to prevent applications from overriding HostnameVerifier
Add allowHttpsHostnameVerifierOverride manifest attribute with a default
of true. When the attribute is false, the platform will always use the
platform default HostnameVerifier even when the application attempts to
use a different HostnameVerifier. This is a best effort mechanism
that is only effective for applications that use the Android built-in
library rather than bringing along their own, although third-party
implementations are encouraged to also honor this flag.

We use a flags2 variable that we added to ApplicationInfo since flags is almost
fully allocated but are open to any other strategies to store the manifest
attribute value.

As described by numerous sources including [1] [2] [3] and [4],
use of an insecure HostnameVerifier is a common source of security vulnerabilities.

This change goes along with a corresponding change to libcore with
the same Change-Id.

[1] Tendulkar and Enck. An Application Package Configuration
Approach to Mitigating Android SSL Vulnerabilities. MoST 2014.

[2] Fahl, et al. Why Eve and Mallory Love Android: An Analysis of
Android SSL (In)Security. CCS '12.

[3] Sounthiraraj, et al. Large Scale, Automated Detection of SSL/TLS
Man-in-the-Middle Vulnerabilities in Android Apps. NDSS '14.

[4] FireEye. SSL Vulnerabilities: Who listens when Android
applications talk?
https://www.fireeye.com/blog/threat-research/2014/08/ssl-vulnerabilities-who-listens-when-android-applications-talk.html

Change-Id: I6df0237576ab9d6c3210960f7a04efae3492e30f