Change the way the connect timeout test works
URLConnectionTest.testConnectTimeouts():

The upcoming version of OkHttp breaks this test.
When I investigated I found it's because the code is
creating a URL like:

http://:::<port number>/

because the stuck server returns "::" as the host
address for the wildcard binding done. The new version
of OkHttp rejects this URL as invalid (as it is missing
[] around the IPv6 address).

The test was supposed to check multiple IPs are tried
and the overall timeout is the sum of all the sockets
tried.

Switching StuckServer over to binding to localhost
did not work.

Since the format of the URL does not appear to be the
purpose of this test it is reasonable to change it
before the OkHttp update.

Instead, the test now replaces the the SocketFactory
with one that returns sockets that immediately time
out, avoid the need to nerf the server-side entirely.

This change adds a new dependency on Mockito so that a
partial mock for Socket can be created.

Bug: 23553256
Change-Id: Iaa57987c59cd6ec1c117f9c90f6ff8ea42d58210